+++ src/main/java/matveyeva/chat/menu/Menu.java
@@ -1,5 +1,7 @@
 package matveyeva.chat.menu;
 
+import matveyeva.chat.Entity.User;
+
 public interface Menu {
-    void showMenu();
+    void showMenu(User user);
 }
