+++ src/main/java/matveyeva/chat/menu/RoomChangesMenu.java
@@ -1,0 +1,56 @@
+package matveyeva.chat.menu;
+
+import java.io.BufferedReader;
+import java.io.BufferedWriter;
+import matveyeva.chat.Entity.User;
+import matveyeva.chat.service.RoomChangesService;
+import org.apache.log4j.Logger;
+
+public class RoomChangesMenu implements Menu {
+
+    protected BufferedReader input;
+    protected BufferedWriter output;
+    private static final org.apache.log4j.Logger logger = Logger.getLogger(RoomChangesMenu.class);
+
+    private RoomChangesService roomChangesService = RoomChangesService.getInstance();
+
+    public RoomChangesMenu(BufferedWriter output, BufferedReader input) {
+        this.output = output;
+        this.input = input;
+    }
+
+    @Override
+    public void showMenu(User user) {
+        boolean check = false;
+        try {
+            while (!check) {
+                roomChangesService
+                    .send("delete room | update room | see all rooms | create room |exit", output);
+                String answer = input.readLine();
+                switch (Integer.parseInt(answer)) {
+                    case 1:
+                        roomChangesService.deleteRoom(output, input, user);
+                        break;
+                    case 2:
+                        roomChangesService.updateRoom(output, input, user);
+                        break;
+                    case 3:
+                        roomChangesService.showAllRooms(output, input, user);
+                        break;
+                    case 4:
+                        roomChangesService.createRoom(output, input, user);
+                        break;
+                    case 5:
+                        check = true;
+                        break;
+                }
+
+            }
+        } catch (NumberFormatException ex) {
+
+        } catch (Exception e) {
+            e.printStackTrace();
+            roomChangesService.send("Something went wrong", output);
+        }
+    }
+}
